moving to share/p2p/shrexeds
