+++ share/p2p/shrexeds/exchange_test.go
@@ -1,4 +1,4 @@
-package p2p
+package shrexeds
 
 import (
 	"context"
