+++ share/p2p/shrexeds/p2p.go
@@ -1,4 +1,4 @@
-package p2p
+package shrexeds
 
 import (
 	"fmt"
