+++ share/p2p/shrexeds/server.go
@@ -1,4 +1,4 @@
-package p2p
+package shrexeds
 
 import (
 	"context"
@@ -12,15 +12,19 @@ import (
 
 	"github.com/celestiaorg/celestia-node/share"
 	"github.com/celestiaorg/celestia-node/share/eds"
-	p2p_pb "github.com/celestiaorg/celestia-node/share/eds/p2p/pb"
+	p2p_pb "github.com/celestiaorg/celestia-node/share/p2p/shrexeds/pb"
 	"github.com/celestiaorg/go-libp2p-messenger/serde"
 )
 
 const (
 	// writeDeadline sets timeout for sending messages to the stream
 	writeDeadline = time.Second * 5
-	// readDeadline sets timeout for reading messages from the stream
+	// readDeadline sets timeout for reading a CAR file from disk
 	readDeadline = time.Minute
+	// getCarDeadline sets timeout for reading a CAR file from disk
+	getCarDeadline = time.Minute
+	// bufferSize is the size of the buffer used for writing an ODS to the stream
+	bufferSize = 32 * 1024
 )
 
 // Server is responsible for serving ODSs for blocksync over the ShrEx/EDS protocol.
@@ -74,7 +78,7 @@ func (s *Server) handleStream(stream network.Stream) {
 		return
 	}
 
-	ctx, cancel := context.WithTimeout(s.ctx, readDeadline)
+	ctx, cancel := context.WithTimeout(s.ctx, getCarDeadline)
 	defer cancel()
 	status := p2p_pb.Status_OK
 	// determine whether the EDS is available in our store
@@ -152,7 +156,8 @@ func (s *Server) writeODS(edsReader io.ReadCloser, stream network.Stream) error 
 	if err != nil {
 		return fmt.Errorf("creating ODS reader: %w", err)
 	}
-	_, err = io.Copy(stream, odsReader)
+	buf := make([]byte, bufferSize)
+	_, err = io.CopyBuffer(stream, odsReader, buf)
 	if err != nil {
 		return fmt.Errorf("writing ODS bytes: %w", err)
 	}
